Added stableswap pallet to testing runtime and moved accountIdFor to common runtime
